feat: Add SQLAlchemy and pymysql dependencies
This commit adds the SQLAlchemy and pymysql dependencies to the project. SQLAlchemy is a powerful SQL toolkit and Object-Relational Mapping (ORM) library, while pymysql is a Python MySQL client library. These dependencies will be used to interact with the database in a more efficient and flexible way.

By adding these dependencies, we can leverage the features provided by SQLAlchemy, such as the ability to define database models using classes and perform CRUD operations using high-level APIs. This will improve the overall functionality and maintainability of the backend application.

These changes enhance the project by incorporating industry-standard tools for working with databases and improving the database interaction capabilities of the application.